+++ packages/mgd/src/index.ts
@@ -10,6 +10,7 @@ import { registerDistCommands } from "./commands/dist.js";
 import { registerApiCommands } from "./commands/api.js";
 import { registerFileCommands } from "./commands/file.js";
 import { registerAspectCommands } from "./commands/aspect.js";
+import { registerSkillsCommands } from "./commands/skills.js";
 
 const program = new Command();
 
@@ -26,6 +27,7 @@ registerDistCommands(program);
 registerApiCommands(program);
 registerFileCommands(program);
 registerAspectCommands(program);
+registerSkillsCommands(program);
 
 try {
     await program.parseAsync(process.argv);